fix: moving Community Watermark to  `ee` folder
