+++ apps/meteor/client/sidebar/footer/voip/index.tsx
@@ -2,6 +2,7 @@ import { useEndpoint, useTranslation } from '@rocket.chat/ui-contexts';
 import React, { ReactElement, useCallback, useMemo, useState } from 'react';
 
 import { useHasLicenseModule } from '../../../../ee/client/hooks/useHasLicenseModule';
+import { SidebarFooterWatermark } from '../../../../ee/client/sidebar/footer/SidebarFooterWatermark';
 import {
 	useCallActions,
 	useCallCreateRoom,
@@ -12,7 +13,6 @@ import {
 	useQueueName,
 } from '../../../contexts/CallContext';
 import SidebarFooterDefault from '../SidebarFooterDefault';
-import { SidebarFooterWatermark } from '../SidebarFooterWatermark';
 import { VoipFooter as VoipFooterComponent } from './VoipFooter';
 
 export const VoipFooter = (): ReactElement | null => {
