+++ apps/meteor/ee/client/sidebar/footer/SidebarFooterWatermark.tsx
